requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.1.0
BaselineRemoval==0.0.8
Brotli==1.0.9
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-extensions==0.1.6
dash-html-components==2.0.0
dash-table==5.0.0
EditorConfig==0.12.3
exceptiongroup==1.0.0rc9
Flask==2.1.2
Flask-Caching==2.0.0
Flask-Compress==1.13
fonttools==4.37.4
future==0.18.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==4.13.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsbeautifier==1.14.6
kiwisolver==1.4.4
macholib==1.16.2
MarkupSafe==2.1.1
matplotlib==3.6.0
more-itertools==8.14.0
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
plotly==5.10.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.5.0
sigfig==1.3.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
